remove create_poll from flex UI, combine PollCard and GroupPollCard widgets
