+++ assets/flexUI.json
@@ -5,7 +5,7 @@
 			"tabbar.default": "favorites",
 
 			"home": [
-				"all_groups", "my_groups", "event_feed", "my_events", "public_surveys", "create_poll", "recent_polls"
+				"all_groups", "my_groups", "event_feed", "my_events", "public_surveys", "recent_polls"
 			],
 			"home.system": ["tout", "connect"],
 			"home.connect": ["phone_or_email"],
@@ -14,7 +14,7 @@
 			"home.groups": ["all_groups", "my_groups"],
 			"home.events": ["event_feed", "my_events"],
 			"home.surveys": ["public_surveys"],
-			"home.polls": ["create_poll", "recent_polls"],
+			"home.polls": ["recent_polls"],
 
 			"home.all_groups": ["groups"],
 			"home.my_groups": ["groups"],
@@ -23,14 +23,13 @@
 			"home.public_surveys": ["surveys"],
 			"home.all_notifications": ["inbox"],
 			"home.unread_notifications": ["inbox"],
-			"home.create_poll": ["polls"],
 			"home.recent_polls": ["polls"],
 
 			"browse": ["groups", "messages", "surveys", "events", "polls"],
 			"browse.events": ["event_feed", "my_events"],
 			"browse.groups": ["all_groups", "my_groups"],
 			"browse.surveys": ["public_surveys"],
-			"browse.polls": ["create_poll", "recent_polls"],
+			"browse.polls": ["recent_polls"],
 
 			"settings": ["favorites", "notifications"],
 			"authenticate": ["connect", "connected", "feedback"],
@@ -189,7 +188,7 @@
 	"defaults": {
 		"favorites": {
 			"content": {
-				"home": ["all_groups", "my_groups", "event_feed", "my_events", "public_surveys", "create_poll", "recent_polls"]
+				"home": ["all_groups", "my_groups", "event_feed", "my_events", "public_surveys", "recent_polls"]
 			},
 			"rules": {
 			}
